Use portable format specifiers (PRIu64/PRId64) for logging

+++ RuntimeManager/AIConfiguration.cpp
@@ -25,6 +25,7 @@
 #include <sstream>
 #include <string>
 #include <climits>
+#include <inttypes.h>
 #include <sys/stat.h>
 #include <yaml-cpp/yaml.h>
 
@@ -368,7 +369,7 @@ namespace Plugin
             if (root["memoryLimit"]) {
                 uint64_t memLimit = root["memoryLimit"].as<uint64_t>();
                 if (memLimit > static_cast<uint64_t>(SSIZE_MAX)) {
-                    LOGERR("memoryLimit value %llu exceeds SSIZE_MAX (%zd), clamping to SSIZE_MAX", memLimit, SSIZE_MAX);
+                    LOGERR("memoryLimit value %" PRIu64 " exceeds SSIZE_MAX (%zd), clamping to SSIZE_MAX", memLimit, SSIZE_MAX);
                     mNonHomeAppMemoryLimit = SSIZE_MAX;
                 } else {
                     mNonHomeAppMemoryLimit = static_cast<ssize_t>(memLimit);
@@ -378,7 +379,7 @@ namespace Plugin
             if (root["gpuMemoryLimit"]) {
                 uint64_t gpuLimit = root["gpuMemoryLimit"].as<uint64_t>();
                 if (gpuLimit > static_cast<uint64_t>(SSIZE_MAX)) {
-                    LOGERR("gpuMemoryLimit value %llu exceeds SSIZE_MAX (%zd), clamping to SSIZE_MAX", gpuLimit, SSIZE_MAX);
+                    LOGERR("gpuMemoryLimit value %" PRIu64 " exceeds SSIZE_MAX (%zd), clamping to SSIZE_MAX", gpuLimit, SSIZE_MAX);
                     mNonHomeAppGpuLimit = SSIZE_MAX;
                 } else {
                     mNonHomeAppGpuLimit = static_cast<ssize_t>(gpuLimit);
@@ -461,10 +462,10 @@ namespace Plugin
                 {
                     long long ramLimit = std::stoll(value);
                     if (ramLimit > SSIZE_MAX) {
-                        LOGERR("ramLimit value %lld exceeds SSIZE_MAX (%zd), clamping to SSIZE_MAX", ramLimit, SSIZE_MAX);
+                        LOGERR("ramLimit value %" PRId64 " exceeds SSIZE_MAX (%zd), clamping to SSIZE_MAX", (int64_t)ramLimit, SSIZE_MAX);
                         mNonHomeAppMemoryLimit = SSIZE_MAX;
                     } else if (ramLimit < 0) {
-                        LOGERR("ramLimit value %lld is negative, setting to 0", ramLimit);
+                        LOGERR("ramLimit value %" PRId64 " is negative, setting to 0", (int64_t)ramLimit);
                         mNonHomeAppMemoryLimit = 0;
                     } else {
                         mNonHomeAppMemoryLimit = static_cast<ssize_t>(ramLimit);
@@ -474,10 +475,10 @@ namespace Plugin
                 {
                     long long gpuLimit = std::stoll(value);
                     if (gpuLimit > SSIZE_MAX) {
-                        LOGERR("gpuMemoryLimit value %lld exceeds SSIZE_MAX (%zd), clamping to SSIZE_MAX", gpuLimit, SSIZE_MAX);
+                        LOGERR("gpuMemoryLimit value %" PRId64 " exceeds SSIZE_MAX (%zd), clamping to SSIZE_MAX", (int64_t)gpuLimit, SSIZE_MAX);
                         mNonHomeAppGpuLimit = SSIZE_MAX;
                     } else if (gpuLimit < 0) {
-                        LOGERR("gpuMemoryLimit value %lld is negative, setting to 0", gpuLimit);
+                        LOGERR("gpuMemoryLimit value %" PRId64 " is negative, setting to 0", (int64_t)gpuLimit);
                         mNonHomeAppGpuLimit = 0;
                     } else {
                         mNonHomeAppGpuLimit = static_cast<ssize_t>(gpuLimit);
